+++ benchmarks/matvec/cuda/matvec_cuda.cu
@@ -43,7 +43,7 @@
 
 #include "fill.hpp"
 
-#include <mdspan.hpp>
+#include <mdspan/mdspan.hpp>
 
 #include <memory>
 #include <random>
