+++ benchmarks/matvec/openmp/matvec_openmp.cpp
@@ -13,7 +13,7 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //@HEADER
-#include <mdspan.hpp>
+#include <mdspan/mdspan.hpp>
 
 #include <memory>
 #include <random>
