+++ benchmarks/sum/cuda/sum_3d_cuda.cu
@@ -41,7 +41,7 @@
 //@HEADER
 */
 
-#include <mdspan.hpp>
+#include <mdspan/mdspan.hpp>
 
 #include <memory>
 #include <random>
