+++ comp_bench/cbench_submdspan.cpp.erb
@@ -1,5 +1,5 @@
 
-#include <mdspan.hpp>
+#include <mdspan/mdspan.hpp>
 
 int test(int* data) {
   #if defined(METABENCH)
