+++ compilation_tests/ctest_extents_ctors.cpp
@@ -15,7 +15,7 @@
 //@HEADER
 #include "ctest_common.hpp"
 
-#include <mdspan.hpp>
+#include <mdspan/mdspan.hpp>
