+++ include/mdspan/experimental/mdarray.hpp
@@ -25,5 +25,5 @@
 #endif
 
 #include "../mdspan.hpp"
-#include "__p1684_bits/mdarray.hpp"
+#include "../../experimental/__p1684_bits/mdarray.hpp"
