+++ tests/test_element_access.cpp
@@ -14,7 +14,7 @@
 //
 //@HEADER
 
-#include <mdspan.hpp>
+#include <mdspan/mdspan.hpp>
 
 #include <gtest/gtest.h>
