+++ tests/test_mdarray_ctors.cpp
@@ -13,7 +13,7 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //@HEADER
-#include <experimental/mdarray.hpp>
+#include <mdspan/experimental/mdarray.hpp>
 #include <vector>
 
 #include <gtest/gtest.h>
