+++ tests/test_submdspan_static_slice.cpp
@@ -13,7 +13,7 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //@HEADER
-#include <mdspan.hpp>
+#include <mdspan/mdspan.hpp>
 #include <cstdint>
 #include <type_traits>
 #include <vector>
